chore simplifying project
