--- projeto-gerar-horario/Microsservicos/GerarHorario-Service/.env.example
@@ -1,4 +1,0 @@
-RABBITMQ_USERNAME=admin
-RABBITMQ_PASSWORD=admin
-
-RABBITMQ_MANAGEMENT_ALLOW_WEB_ACCESS=true
