+++ projeto-gerar-horario/Microsservicos/GerarHorario-Service/GerarHorario-Service.csproj
@@ -1,29 +1,21 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
     <RootNamespace>GerarHorario_Service</RootNamespace>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
-
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\GerarHorario\GerarHorario.csproj" />
-  </ItemGroup>
-
 
   <ItemGroup>
     <PackageReference Include="RabbitMQ.Client" Version="6.8.1" />
-    <PackageReference Include="RabbitMQ.Stream.Client" Version="1.8.2" />
   </ItemGroup>
-
 
   <ItemGroup>
-    <Folder Include="Extensions\" />
+    <Reference Include="GerarHorario">
+      <HintPath>..\..\GerarHorario\bin\Debug\net8.0\GerarHorario.dll</HintPath>
+    </Reference>
   </ItemGroup>
 
-
-  <ItemGroup>
-    <Compile Remove="Extensions\MiddlewareExtensions.cs" />
-  </ItemGroup>
 </Project>
