+++ projeto-gerar-horario/Microsservicos/GerarHorario-Service/Program.cs
@@ -1,23 +1,73 @@
-using GerarHorario_Service.Middlewares;
+﻿using System.Text;
+using System.Text.Json;
+using RabbitMQ.Client;
+using RabbitMQ.Client.Events;
+using Sisgea.GerarHorario;
+using Sisgea.GerarHorario.Core.Dtos.Configuracoes;
+using Sisgea.GerarHorario.Core.Dtos.HorarioGerado;
 
 
-Thread server = new(() =>
+var factory = new ConnectionFactory()
 {
+    HostName = "localhost",
+    UserName = "admin",
+    Password = "admin"
+};
 
-    var builder = WebApplication.CreateBuilder(args);
-    var app = builder.Build();
+using var connection = factory.CreateConnection();
 
-    app.MapGet("/up", () => "up");
+using var channel = connection.CreateModel();
+channel.QueueDeclare(
+    queue: "gerar_horario",
+    durable: true,
+    exclusive: false,
+    autoDelete: false,
+    arguments: null
+);
 
+channel.QueueDeclare(
+    queue: "horario_gerado",
+    durable: true,
+    exclusive: false,
+    autoDelete: false,
+    arguments: null
+);
 
+void publicarRespostaGerarHorario(HorarioGerado? horarioGerado = null)
+{
+    // TODO: message = JSON(horarioGerado)
+    const string message = "Hello World!";
+    var body = Encoding.UTF8.GetBytes(message);
 
-    app.Run();
-});
+    channel.BasicPublish(exchange: string.Empty,
+        routingKey: "horario_gerado",
+        basicProperties: null,
+        body: body);
 
-Thread queue = new(QueueService.ListenQueue);
+    Console.WriteLine($" [x] Sent {message}");
+}
 
+Console.WriteLine(" [*] Waiting for messages.");
 
+var consumer = new EventingBasicConsumer(channel);
+consumer.Received += (model, ea) =>
+{
+    var body = ea.Body.ToArray();
+    var message = Encoding.UTF8.GetString(body);
+    Console.WriteLine($" [x] Received {message}");
 
-server.Start();
-queue.Start();
+    var options = new JsonSerializerOptions()
+    {
+        PropertyNameCaseInsensitive = true
+    };
 
+    GerarHorarioOptions? gerarHorarioOptions = JsonSerializer.Deserialize<GerarHorarioOptions>(message, options);
+    Console.WriteLine(gerarHorarioOptions);
+
+    publicarRespostaGerarHorario();
+};
+
+channel.BasicConsume(queue: "gerar_horario",
+    autoAck: true,
+    consumer: consumer);
+
